+++ src/js/PlotBoilerplate.js
@@ -240,6 +240,10 @@ var PlotBoilerplate = /** @class */ (function () {
                 color: '#a8a8a8',
                 lineWidth: 1
             },
+            selectedVertex: {
+                color: '#c08000',
+                lineWidth: 2
+            },
             line: {
                 color: '#a844a8',
                 lineWidth: 1
@@ -685,7 +689,7 @@ var PlotBoilerplate = /** @class */ (function () {
      * This is just a tiny helper function to determine the render color of vertices.
      **/
     PlotBoilerplate.prototype._handleColor = function (h, color) {
-        return h.attr.draggable ? color : 'rgba(128,128,128,0.5)';
+        return h.attr.isSelected ? this.drawConfig.selectedVertex.color : (h.attr.draggable ? color : 'rgba(128,128,128,0.5)');
     };
     /**
      * Draw all drawables.
@@ -709,11 +713,11 @@ var PlotBoilerplate = /** @class */ (function () {
                     this.draw.cubicBezier(d.bezierCurves[c].startPoint, d.bezierCurves[c].endPoint, d.bezierCurves[c].startControlPoint, d.bezierCurves[c].endControlPoint, this.drawConfig.bezier.color, this.drawConfig.bezier.lineWidth);
                     if (this.drawConfig.drawBezierHandlePoints && this.drawConfig.drawHandlePoints) {
                         if (!d.bezierCurves[c].startPoint.attr.bezierAutoAdjust) {
-                            this.draw.diamondHandle(d.bezierCurves[c].startPoint, 7, this._handleColor(d.bezierCurves[c].startPoint, 'orange'));
+                            this.draw.diamondHandle(d.bezierCurves[c].startPoint, 7, this._handleColor(d.bezierCurves[c].startPoint, this.drawConfig.vertex.color));
                             d.bezierCurves[c].startPoint.attr.renderTime = renderTime;
                         }
                         if (!d.bezierCurves[c].endPoint.attr.bezierAutoAdjust) {
-                            this.draw.diamondHandle(d.bezierCurves[c].endPoint, 7, this._handleColor(d.bezierCurves[c].endPoint, 'orange'));
+                            this.draw.diamondHandle(d.bezierCurves[c].endPoint, 7, this._handleColor(d.bezierCurves[c].endPoint, this.drawConfig.vertex.color));
                             d.bezierCurves[c].endPoint.attr.renderTime = renderTime;
                         }
                         this.draw.circleHandle(d.bezierCurves[c].startControlPoint, 3, this._handleColor(d.bezierCurves[c].startControlPoint, '#008888'));
@@ -838,7 +842,7 @@ var PlotBoilerplate = /** @class */ (function () {
         // Draw all vertices as small squares if they were not already drawn by other objects
         for (var i in this.vertices) {
             if (this.drawConfig.drawVertices && this.vertices[i].attr.renderTime != renderTime) {
-                this.draw.squareHandle(this.vertices[i], 5, this.vertices[i].attr.isSelected ? 'rgba(192,128,0)' : this._handleColor(this.vertices[i], 'rgb(0,128,192)'));
+                this.draw.squareHandle(this.vertices[i], 5, this._handleColor(this.vertices[i], 'rgb(0,128,192)'));
             }
         }
     };