+++ src/ts/interfaces.ts
@@ -114,6 +114,7 @@ export interface DrawConfig {
     ellipse : DrawSettings;
     circle : DrawSettings;
     vertex : DrawSettings;
+    selectedVertex : DrawSettings;
     line : DrawSettings;
     vector : DrawSettings;
     image : DrawSettings;