+++ changelog.md
@@ -4,6 +4,7 @@
 * 2020-07-28
   * Added PlotBoilerplate.revertMousePosition(number,number) –  the inverse function of transformMousePosition(...).
   * Added Bézier curve splitting to the bezier-point-distance demo.
+  * KeyHandler: changed the 'delete' key code from 8 to 46.
 * 2020-07-27
   * Added the getVertexNear(XYCoords,number) function.
   * Extended the remove(Drawable) function: vertices are now removed, too.